requirements.txt
pandas>=1.3.0
requests>=2.25.1
psycopg2-binary>=2.9.0
matplotlib>=3.4.0
pyyaml>=6.0